+++ fs/cifsd/ksmbd_server.h
@@ -30,10 +30,8 @@ struct ksmbd_heartbeat {
  */
 #define KSMBD_GLOBAL_FLAG_INVALID		(0)
 #define KSMBD_GLOBAL_FLAG_SMB2_LEASES		BIT(0)
-#define KSMBD_GLOBAL_FLAG_CACHE_TBUF		BIT(1)
-#define KSMBD_GLOBAL_FLAG_CACHE_RBUF		BIT(2)
-#define KSMBD_GLOBAL_FLAG_SMB2_ENCRYPTION	BIT(3)
-#define KSMBD_GLOBAL_FLAG_SMB3_MULTICHANNEL	BIT(4)
+#define KSMBD_GLOBAL_FLAG_SMB2_ENCRYPTION	BIT(1)
+#define KSMBD_GLOBAL_FLAG_SMB3_MULTICHANNEL	BIT(2)
 
 struct ksmbd_startup_request {
 	__u32	flags;
