+++ fs/cifsd/mgmt/user_session.c
@@ -14,7 +14,6 @@
 #include "tree_connect.h"
 #include "../transport_ipc.h"
 #include "../connection.h"
-#include "../buffer_pool.h"
 #include "../vfs_cache.h"
 
 static DEFINE_IDA(session_ida);
